transmission: Use common view for configuration
